refactor: rename ObjectField.story.vue to object-field.story.vue for consistency
